+++ interop/ksp/test-processor/src/main/kotlin/com/squareup/kotlinpoet/ksp/test/processor/exampleAnnotations.kt
@@ -86,4 +86,6 @@ enum class AnnotationEnumValue {
 
 annotation class AnnotationWithVararg(val simpleArg: Int, vararg val args: String)
 
+annotation class AnnotationWithVarargFirst(vararg val args: String, val simpleArg: Int)
+
 annotation class AnnotationWithTypeArgs<T, R>